# Conflicts:
#	src/test/java/de/uni_passau/fim/se2/litterbox/jsoncreation/JSONFileCreatorTest.java
#	src/test/java/de/uni_passau/fim/se2/litterbox/jsoncreation/JSONStringCreatorTest.java

+++ src/main/java/de/uni_passau/fim/se2/litterbox/ast/parser/NumExprParser.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2020 LitterBox contributors
+ * Copyright (C) 2019-2021 LitterBox contributors
  *
  * This file is part of LitterBox.
  *
@@ -278,7 +278,7 @@ public class NumExprParser {
     static NumberLiteral parseNumber(JsonNode inputs, String inputKey) throws ParsingException {
         Preconditions.checkArgument(inputs.has(inputKey));
         String valueString = ExpressionParser.getDataArrayByName(inputs, inputKey).get(POS_INPUT_VALUE).asText();
-        float value = Float.parseFloat(valueString);
+        double value = Double.parseDouble(valueString);
         return new NumberLiteral(value);
     }
